[ADD] Add some variables

+++ rds/tests/test.ps1
@@ -7,10 +7,20 @@ $tests | ForEach-Object {
   $TF_VARIABLES | ConvertTo-Json -Depth 100
   ce terraform plan -no-color -input=false `
     -var-file "vars/${_}.json"
-  ce terraform apply -auto-approve -no-color -input=false `
-    -var-file "vars/${_}.json"
-  ce terraform destroy -auto-approve -no-color -input=false `
-    -var-file "vars/${_}.json"
+
+  if ($TF_APPLY -ieq "true")
+  {
+    try
+    {
+      ce terraform apply -auto-approve -no-color -input=false `
+        -var-file "vars/${_}.json"
+    }
+    finally
+    {
+      ce terraform destroy -auto-approve -no-color -input=false `
+        -var-file "vars/${_}.json"
+    }
+  }
 }
